requirements.txt
Flask~=2.1.2
Jinja2~=3.1.2
MarkupSafe~=2.1.1
Werkzeug~=1.0.1
click~=8.1.3
colorama~=0.4.5
importlib-metadata~=4.11.4
itsdangerous~=2.1.2
pip~=20.2.3
pymongo~=4.1.1
python-dotenv~=0.19.2
setuptools~=49.2.1
zipp~=3.8.0
Flask-JWT-Extended~=4.4.1
gunicorn~=20.1.0
